src/modules/auth/auth_module.py
from nest.core import Module
from .auth_controller import AuthController
from .auth_service import AuthService


@Module(
    controllers=[AuthController],
    providers=[AuthService],
    imports=[]
)   
class AuthModule:
    pass

    