requirements.txt
pynest-api
uvicorn[standard]
supabase
pydantic[email]
python-dotenv
